created PeopleTable component and realized all the changeds necessaries
